fix(i18n): client owns renderTemplate; drop languages /runtime dep (PR #1527 CI)
The downstream SvelteKit build failed: Provider → /i18n → context.svelte.ts
imported @happyvertical/smrt-languages/runtime, so every smrt-svelte consumer
(incl. bundled apps) had to resolve the languages package — but it is only an
optional peer with a heavy server dep tree (smrt-core/sql/ai/jobs), so Rollup
could not resolve it.

Fix: the client owns a tiny dependency-free renderTemplate (src/i18n/render.ts);
the languages root is now imported only by the Node-only /i18n/server subpath.
render-parity.spec.ts pins the client copy against the canonical languages
implementation so the {var} contract never drifts. Reverted the speculative
languages /runtime subpath (utils/types/vite/exports/AGENTS) and its vitest
alias. ADR + AGENTS updated to reflect the client-owned helper.

Validation: full build 50/50 (client context bundle has 0 languages refs),
turbo typecheck 78/78 (a11y gate clean), 22 i18n tests green incl parity.
