+++ docs/content/architecture/i18n.md
@@ -66,21 +66,24 @@ or a user name never has to round-trip to the server.
 
 ### Interpolation contract
 
-Single-brace placeholders: `"Showing {count} of {total}"`. This is exactly
+Single-brace placeholders: `"Showing {count} of {total}"`. This matches
 `@happyvertical/smrt-languages`'s `renderTemplate` (`/\{([^{}]+)\}/g`): `Date` →
-ISO string, array/object → JSON, `null` / `undefined` → empty string. The client
-imports `renderTemplate` from the **browser-safe `/runtime` subpath** (below) so
-there is no second implementation to drift.
+ISO string, array/object → JSON, `null` / `undefined` → empty string.
 
-### `@happyvertical/smrt-languages/runtime` (new, browser-safe)
+### The client owns interpolation (`src/i18n/render.ts`)
 
 The languages package root drags `smrt-core`, `sql`, `ai`, `smrt-jobs`, … — fine
-on the server, wrong for a browser bundle. Following the package's existing
-subpath split (root = read path, `/jobs` = worker, `/cli` = admin), a new
-**`/runtime`** subpath exports only the pure, dependency-free helpers
-(`renderTemplate`, `normalizeLocale`, `buildLocaleFallbackChain`) with no
-`node:*` imports. The client i18n layer imports from `/runtime`; the server
-snapshot helper imports the root.
+on the server, wrong for a browser bundle. And because the always-bundled client
+path (`Provider` → `/i18n`) reaches the interpolation helper, importing the
+languages package there would force **every** smrt-svelte consumer (including
+downstream SvelteKit apps) to install that heavy tree — even though languages is
+only an optional peer needed for the server bridge.
+
+So the client owns a tiny, dependency-free copy of `renderTemplate`
+(`src/i18n/render.ts`). The contract is small and stable;
+`__tests__/render-parity.spec.ts` pins the copy against the canonical languages
+implementation so the two never drift. Only the **server** subpath
+(`/i18n/server`) imports the languages root.
 
 ### Key naming
 
@@ -155,6 +158,8 @@ flip to strict as their extraction completes (mirrors the design-token ratchets)
   auto-translation are the reason the languages package exists.
 - **Client async resolution (`await` per string).** Keeps everything dynamic but
   forces loading UI into every label — awkward and a11y-noisy. Rejected.
-- **Re-implement `renderTemplate` in smrt-svelte.** Avoids touching languages
-  but duplicates the interpolation contract (drift risk). Rejected in favour of
-  the `/runtime` subpath (single source of truth).
+- **Import a browser-safe `/runtime` subpath from languages on the client.**
+  Cleanest "single source of truth", but the always-bundled client path reaches
+  it, so it would make the otherwise-optional languages package (and its heavy
+  server dep tree) a hard requirement for every smrt-svelte consumer. Rejected:
+  the client owns a tiny `renderTemplate` copy guarded by a parity test instead.
