+++ packages/languages/src/types.ts
@@ -1,5 +1,4 @@
 import type { SmrtClassOptions } from '@happyvertical/smrt-core';
-import type { LanguageVariables } from './runtime.js';
 
 /** A registered code default for a single (key, locale) pair. */
 export interface LanguageStringDefinitionInput {
@@ -35,7 +34,7 @@ export interface LanguagesPackageConfig {
   [key: string]: unknown;
 }
 
-export type { LanguageVariables };
+export type LanguageVariables = Record<string, unknown>;
 
 export interface ResolveLanguageStringOptions {
   db?: SmrtClassOptions['db'];
